method is ready

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
@@ -19,5 +19,5 @@ public interface UserService {
 
     List<Role> getAllRoles();
 
-    Set<Role> findRolesById (Integer[] RoleIds);
+    Set<Role> findRolesById (Long[] RoleIds);
 }
